join & login

+++ controller/task.controller.js
@@ -4,6 +4,7 @@ import Task from "../model/Task.js";
 export const postTask = async (req, res) => {
   try {
     const { task, isComplete } = req.body;
+    const { userId } = req;
     // const newTask = new Task({ task, isComplete });
     //  await newTask.save();
 
@@ -11,6 +12,7 @@ export const postTask = async (req, res) => {
     const newTask = await Task.create({
       task,
       isComplete,
+      author: userId,
     });
     res.status(200).json({ status: "ok", data: newTask });
   } catch (err) {
@@ -20,7 +22,7 @@ export const postTask = async (req, res) => {
 
 export const getTask = async (req, res) => {
   try {
-    const taskList = await Task.find({}).select("-__v");
+    const taskList = await Task.find({}).select("-__v").populate("author");
     res.status(200).json({ status: "ok", data: taskList });
   } catch (err) {
     res.status(400).json({ status: "fail", error: err });
@@ -33,6 +35,12 @@ export const putTask = async (req, res) => {
     if (!task) {
       throw new Error("App can not find the task");
     }
+
+    // 권한 체크
+    if (task.author.toString() !== req.userId) {
+      return res.status(403).json({ message: "변경 권한이 없습니다." });
+    }
+
     const fields = Object.keys(req.body);
     fields.map((item) => (task[item] = req.body[item]));
     await task.save();
@@ -44,6 +52,12 @@ export const putTask = async (req, res) => {
 
 export const deleteTask = async (req, res) => {
   try {
+    const task = await Task.findById(req.params.id);
+    console.log("task", task.auth);
+    if (task.author.toString() !== req.userId) {
+      return res.status(403).json({ message: "삭제 권한이 없습니다." });
+    }
+    console.log();
     const deleteItem = await Task.findByIdAndDelete(req.params.id);
     res.status(200).json({ status: "ok", data: deleteItem });
   } catch (err) {
